rename and move component
